글쓰기 2

+++ src/main/java/com/ll/exam/app10/app/fileUpload/entity/GenFile.java
@@ -14,15 +14,15 @@ import javax.persistence.Entity;
 @SuperBuilder
 @ToString(callSuper = true)
 public class GenFile extends BaseEntity {
-    private String relTypeCode;
-    private int relId;
-    private String typeCode;
-    private String type2Code;
-    private String fileExtTypeCode;
-    private String fileExtType2Code;
-    private int fileSize;
-    private int fileNo;
-    private String fileExt;
-    private String fileDir;
-    private String originFileName;
+    private String relTypeCode; // 어느것과 과련되었냐
+    private long relId; // 어떤것과 관련 되어 있는데 그 Id는 어떤 것이냐
+    private String typeCode; // 어디에 쓰일 것이냐 ( 용도 )
+    private String type2Code; // 어디에 쓰일 것이냐 2 ( 용도 2 )
+    private String fileExtTypeCode; // 파일 종류 단계 ( img )
+    private String fileExtType2Code; // 파일 종류 단계 ( jpg )
+    private int fileSize; // 파일 크기
+    private int fileNo; // 파일 번호
+    private String fileExt; // 파일 확장자
+    private String fileDir; // 파일 경로
+    private String originFileName; // 원본 파일 이름
 }
